Revert "JFB-2"
